requirements.txt
Flask==2.0.2
google-cloud-speech==2.12.0
google-cloud-texttospeech==2.10.0
requests==2.26.0
twilio==7.0.0
google-auth==2.0.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.22.0
Werkzeug==2.0.3
